+++ lib/src/map_extensions.dart
@@ -5,11 +5,4 @@
 extension MapExtensions<K, V> on Map<K, V> {
   /// Like [Map.entries], but returns each entry as a record.
   Iterable<(K, V)> get pairs => entries.map((e) => (e.key, e.value));
-
-  /// Like [Map.addEntries], but takes each entry as a record.
-  void addPairs(Iterable<(K, V)> newPairs) {
-    for (var (key, value) in newPairs) {
-      this[key] = value;
-    }
-  }
 }