+++ CHANGELOG.md
@@ -5,6 +5,7 @@
     - `copy`: copies an instance without recalculating the canonical values of the keys.
     - `toMap`: creates a `Map<K,V>` (with the original key values).
     - `toMapOfCanonicalKeys`: creates a `Map<C,V>` (with the canonicalized keys).
+- Fixes bugs in `ListSlice.slice` and `ListExtensions.slice`.
 - lints: ^2.0.1
 
 ## 1.17.2